FP16->BF16

+++ include/activation_pool.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "config.h"
-#include <cuda_fp16.h>
+#include <cuda_bf16.h>
 
 // Pre-allocated GPU memory pool for intermediate activations.
 // Uses ping-pong strategy: layer N writes to buf_a, layer N+1 reads from buf_a
@@ -15,19 +15,19 @@ public:
     void init(const ModelConfig& config);
 
     // Get current read/write buffers (alternate each layer)
-    half* input_buf()  { return (flip_ ? buf_b_ : buf_a_); }
-    half* output_buf() { return (flip_ ? buf_a_ : buf_b_); }
+    __nv_bfloat16* input_buf()  { return (flip_ ? buf_b_ : buf_a_); }
+    __nv_bfloat16* output_buf() { return (flip_ ? buf_a_ : buf_b_); }
     void  flip()       { flip_ = !flip_; }
     void  reset()      { flip_ = false; }
 
     // Scratch space for attention scores, logits, etc.
-    half*  scratch()   { return scratch_; }
+    __nv_bfloat16*  scratch()   { return scratch_; }
     float* logits()    { return logits_; }   // [vocab_size] in fp32
 
 private:
-    half*  buf_a_    = nullptr;
-    half*  buf_b_    = nullptr;
-    half*  scratch_  = nullptr;
+    __nv_bfloat16*  buf_a_    = nullptr;
+    __nv_bfloat16*  buf_b_    = nullptr;
+    __nv_bfloat16*  scratch_  = nullptr;
     float* logits_   = nullptr;   // fp32 for numerical stability
     bool   flip_     = false;
 };
